chore: rename folder to persistence for jpa
